requirements.txt
openai==0.26.5
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1
